Update receiver.c
Added a few functions at the end. Added a few define variables for good measure. Adjusted the MAXPULSES devided by 8 of the original. Added a voidFuncPtr but I have no idea what I was doing with that. Added to listenSignal to check if the pulses are of an acceptable amount, otherwise set playerNumber to 0. Fixed some of the function names to be "receiver_*". Please verify.