Moved some of the game related logic to game.c, it was originally in communication.c. Now the enemies are put in an array and then they're added to the enemy library from game.c. Added some defines to player lib to denote the maximum allowed and minimum allowed player numbers. Added a TO-DO list to communication.c that's related to both the communication and game libraries.
communication.c specific:
Updated one of the responses so that now it sends the player number.
Added some safety escape condition to getNext.
Added a bit of documentation on top of each different type of parsing function.
Added some minimum/maximum length exit conditions to these parsing functions.
Created a resetBuffer function since the code was being used in more than one place.
Added a condition that after the string type it must have a comma, and before the final hashtag it must also have a comma; if this is not met then it returns.
Made it so that endTag must be larger than startTag and they must both be positive.

Signed-off-by: Cesar Andreu <cesarandreu@gmail.com>